fix: call the correct URL to get artifact properties
URL called was hardcoded as GET artifact_path?properties\[=x[,y]], which is
interpreted by Artifactory as getting ?properties\[ with a value of
x,[y]], which in turn only returns the file/folder information (like if
you called GET artifact_path). Artifactory API is weird, here if you
call it with a bad parameter it's the same as no parameter.
This fixes #50

Tests are still broken, because other methods relied on the (wrong)
behaviour of properties(). I will implement a get() to have this
information available again.