Correct member-of-groups usage for access tokens
The documentation suggests that group names can be ", " delimited.
This works for group names that does not contain spaces.

Group names with spaces are supported.
To support this we must quote the group list, as `pyartifactory` does.
This also requires us to change group delimiter to ",".
I would assume this is due to Artifactory not being able to destinguish
between the delimiter space and any potential group name spaces.

Correct by using "," in combination with quoting the list.